requirements.txt.txt
flask
sqlalchemy
mysql-connector-python
pymysql